chore: rename streaming swaps -> DCA
