+++ bouncer/tests/DCA_test.ts
@@ -1,5 +1,5 @@
 #!/usr/bin/env -S pnpm tsx
-import { testDCASwaps } from '../shared/streaming_swaps';
+import { testDCASwaps } from '../shared/DCA_test';
 import { executeWithTimeout } from '../shared/utils';
 
 await executeWithTimeout(testDCASwaps(), 150);
