+++ bouncer/tests/all_concurrent_tests.ts
@@ -11,7 +11,7 @@ import { testBrokerFeeCollection } from '../shared/broker_fee_collection';
 import { testBoostingSwap } from '../shared/boost';
 import { observeBadEvent } from '../shared/utils/substrate';
 import { testFillOrKill } from '../shared/fill_or_kill';
-import { testDCASwaps } from '../shared/streaming_swaps';
+import { testDCASwaps } from '../shared/DCA_test';
 
 const swapContext = new SwapContext();
